pyproject.toml
[project]
name = "mensara"
version = "0.1.0"
description = ""
authors = [
    {name = "Scott.McKenzie.Lewis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "typer (>=0.24.1,<0.25.0)",
    "click (>=8.3.1,<9.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.4,<0.16.0)",
    "pylint (>=4.0.5,<5.0.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
